Move new pages out of new/.
