--- src/pages/donate.js
@@ -1,9 +1,0 @@
-import React from "react";
-import DonateView from "../legacy/components/pages/donate/DonateView";
-import Layout from "../legacy/layouts/Layout";
-
-export default () => (
-  <Layout>
-    <DonateView />
-  </Layout>
-);
