+++ src/pages/volunteer/index.js
@@ -1,20 +1,20 @@
 import React, { useState } from "react";
 import { Helmet } from "react-helmet";
 
-import ArticleSpotlightCard from "../../../components/grid-aware/ArticleSpotlightCard";
-import COVID19infoBoxBlock from "../../../components/grid-aware/COVID19InfoBoxBlock";
-import ImageHeader from "../../../components/grid-aware/ImageHeader";
-import classroom from "../../../components/grid-aware/ImageHeader/stories/classroom.png";
-import whiteboard from "../../../components/grid-aware/ImageHeader/stories/whiteboard-sticky-notes.png";
-import Modal from "../../../components/grid-aware/Modal";
-import OneParagraphBlock from "../../../components/grid-aware/OneParagraphBlock";
-import Spacer from "../../../components/grid-aware/Spacer";
-import StatsBlock from "../../../components/grid-aware/StatsBlock";
-import ThreeParagraphBlock from "../../../components/grid-aware/ThreeParagraphBlock";
-import TwoParagraphBlock from "../../../components/grid-aware/TwoParagraphBlock";
-import volunteers from "../../../components/grid-aware/TwoParagraphBlock/stories/volunteers.png";
-import Layout from "../../../components/layout";
-import VolunteerSignupForm from "../../../components/thirdparty/mailchimp/VolunteerSignupForm";
+import ArticleSpotlightCard from "../../components/grid-aware/ArticleSpotlightCard";
+import COVID19infoBoxBlock from "../../components/grid-aware/COVID19InfoBoxBlock";
+import ImageHeader from "../../components/grid-aware/ImageHeader";
+import classroom from "../../components/grid-aware/ImageHeader/stories/classroom.png";
+import whiteboard from "../../components/grid-aware/ImageHeader/stories/whiteboard-sticky-notes.png";
+import Modal from "../../components/grid-aware/Modal";
+import OneParagraphBlock from "../../components/grid-aware/OneParagraphBlock";
+import Spacer from "../../components/grid-aware/Spacer";
+import StatsBlock from "../../components/grid-aware/StatsBlock";
+import ThreeParagraphBlock from "../../components/grid-aware/ThreeParagraphBlock";
+import TwoParagraphBlock from "../../components/grid-aware/TwoParagraphBlock";
+import volunteers from "../../components/grid-aware/TwoParagraphBlock/stories/volunteers.png";
+import Layout from "../../components/layout";
+import VolunteerSignupForm from "../../components/thirdparty/mailchimp/VolunteerSignupForm";
 import teamActivityImage from "./image4.png";
 import pairProgrammingImage from "./image5.png";
 import teamSnackingImage from "./image6.png";
